refactor(indexer): one apply path, with leading as two named hooks (#5917)
Three complete paths from a replica-log record to local state — a following one, a leading one and a third for the catch-up in between — become one, with leading reaching it through a fixed pair of hooks and nowhere else. The three agreed on every record kind already; what they did not share was the code, so each fence check, staleness rule and block hold existed three times and had to be kept in step by hand.

- **`Leadership` is the strategy, and enumerating its hooks is the point**
  `applyAuthored` answers for a record this node produced, from what it still holds. `takeCut` writes out the cut for a boundary it wrote. Two, because there are exactly two things the producer of a record knows that a reader of it does not, and naming them is what makes a third a hole rather than an addition.

  - **A question asked of something absent, rather than a branch on the role**
    With no leadership both hooks have an ordinary answer, so a new kind of replica message needs no decision at all unless it wants a hook of its own — and wanting one is then visible in the interface rather than buried in a `when`.

  - **The cache is an optimisation; the cut is a responsibility**
    `applyAuthored` is safe only because a record and what leadership holds are the same thing by construction: it appends in resolution order and reads back in position order, so a hit whose contents disagree is a fault rather than a case to handle. `takeCut` is different in kind — a cut is local state as of exactly one position, so the boundary's author is the only node whose state that is, and it is the one effect that could not have been done when the record was produced.

- **The fence subsumes the leader's own term comparison rather than sitting beside it**
  The leader discarded records below its term itself, having already been told by the partition fence. Those are the same verdict: the fence's highest is at or above the leader's term the moment its claim is read back, so `admit` returning false is exactly "below our term". What remains genuinely the leader's is supersession, which the applier raises from the term the fence just admitted.

- **The catch-up that finishes a predecessor's block is leadership with one hook**
  It authored nothing, so it answers for no record; it has read every record below the boundary, so its state at that cut is the state the predecessor would have written, and it can take the cut. That is the whole of what the deleted processor did differently — and expressing it as a leadership rather than a third apply path is what let the class go.

- **Observability follows the reader rather than the role**
  A leader now records the same per-message-type timings and block-hold measurements a follower does, under the same names and the same `db` tag. A measurement that restarts when the role changes cannot tell a gap apart from a node that has stopped reading; attached to the reader, the same numbers answer both.

Behaviour-preserving, with one deliberate widening: the staleness check now also applies on the leading path, where previously there was none. It can only skip work, and it is asked after the first hook — so a record this node has only just written is never treated as a replay of something it already had, which is the one case where a watermark comparison would say otherwise.

Refs #5917
