--- core/src/main/kotlin/xtdb/indexer/TransitionLogProcessor.kt
@@ -1,163 +1,0 @@
-package xtdb.indexer
-
-import org.apache.arrow.memory.BufferAllocator
-import xtdb.api.DatabaseName
-import xtdb.api.TransactionKey
-import xtdb.api.TransactionResult
-import xtdb.api.log.DbOp
-import xtdb.api.log.Log
-import xtdb.types.MessageId
-import xtdb.api.log.ReplicaMessage
-import xtdb.api.log.Watchers
-import xtdb.api.storage.Storage
-import xtdb.api.error.Anomaly
-import xtdb.api.error.Interrupted
-import xtdb.database.Database
-import xtdb.database.PartitionState
-import xtdb.storage.BufferPool
-import xtdb.table.fromSchemaAndTable
-import xtdb.util.*
-import xtdb.util.StringUtil.asLexHex
-
-private val LOG = TransitionLogProcessor::class.logger
-
-// State, not a Service: driven synchronously through the transition, no loop of its own — so
-// it has no term scope to cancel, just `use` it.
-class TransitionLogProcessor(
-    allocator: BufferAllocator,
-    private val bufferPool: BufferPool,
-    private val partitionState: PartitionState,
-    private val dbName: DatabaseName,
-    private val liveIndex: LiveIndex,
-    private val blockUploader: BlockUploader,
-    private val watchers: Watchers,
-    private val dbCatalog: Database.Catalog?,
-    afterReplicaMsgId: MessageId,
-    private val hasExternalSource: Boolean,
-    private val termId: Long = 0,
-) : LogProcessor.Processor<ReplicaMessage> {
-
-    override var latestReplicaMsgId: MessageId = afterReplicaMsgId
-        private set
-
-    private val blockCatalog = partitionState.blockCatalog
-    private val trieCatalog = partitionState.trieCatalog
-
-    private val allocator = allocator.newChildAllocator("transition-log-processor", 0, Long.MAX_VALUE)
-
-    private val ReplicaMessage.stale get() =
-        when (this) {
-            is ReplicaMessage.ResolvedTx -> txId <= watchers.latestTxId
-            is ReplicaMessage.TriesAdded -> sourceMsgId <= watchers.latestSourceMsgId
-            is ReplicaMessage.BlockBoundary -> blockIndex <= (blockCatalog.currentBlockIndex ?: -1)
-            is ReplicaMessage.BlockUploaded -> blockIndex <= (blockCatalog.currentBlockIndex ?: -1)
-            is ReplicaMessage.NoOp -> srcMsgId != null && srcMsgId <= watchers.latestSourceMsgId
-            is ReplicaMessage.TriesDeleted -> false
-        }
-
-    private suspend fun processRecord(record: Log.Record<ReplicaMessage>) {
-        val msgId = record.msgId
-        when (val msg = record.message) {
-            is ReplicaMessage.ResolvedTx -> {
-                val txKey = TransactionKey(msg.txId, msg.systemTime)
-
-                val tables = msg.loadTableData(allocator)
-                try {
-                    liveIndex.commitTx(txKey, tables)
-                } finally {
-                    tables.closeAll()
-                }
-
-                if (msg.committed) {
-                    when (val dbOp = msg.dbOp) {
-                        is DbOp.Attach -> if (dbCatalog != null) {
-                            try {
-                                dbCatalog.attach(dbOp.dbName, dbOp.config)
-                            } catch (e: Anomaly.Caller) {
-                                LOG.debug(e) { "[$dbName] transition: attach database '${dbOp.dbName}' failed" }
-                            }
-                        }
-                        is DbOp.Detach -> if (dbCatalog != null) {
-                            try {
-                                dbCatalog.detach(dbOp.dbName)
-                            } catch (e: Anomaly.Caller) {
-                                LOG.debug(e) { "[$dbName] transition: detach database '${dbOp.dbName}' failed" }
-                            }
-                        }
-                        null -> {}
-                    }
-                }
-
-                val result =
-                    if (msg.committed) TransactionResult.Committed(txKey)
-                    else TransactionResult.Aborted(txKey, msg.error)
-
-                // Handling for pre-`f3eb8d7d9` ResolvedTx records — see #5586.
-                val effectiveSrcMsgId = msg.srcMsgId
-                    ?: if (hasExternalSource) watchers.latestSourceMsgId else msg.txId
-                watchers.notifyTx(result, effectiveSrcMsgId, msg.externalSourceToken)
-            }
-
-            is ReplicaMessage.TriesAdded -> {
-                if (msg.storageVersion == Storage.VERSION && msg.storageEpoch == bufferPool.epoch) {
-                    msg.tries.groupBy { it.tableName }.forEach { (tableName, tries) ->
-                        trieCatalog.addTries(
-                            fromSchemaAndTable(tableName),
-                            tries,
-                            record.logTimestamp
-                        )
-                    }
-                }
-                watchers.notifyMsg(msg.sourceMsgId)
-            }
-
-            is ReplicaMessage.BlockBoundary -> {
-                LOG.debug("[$dbName] block boundary b${msg.blockIndex.asLexHex}: source=${msg.latestProcessedMsgId}, replica=$msgId")
-                blockUploader.uploadBlock(msgId, termId, msg)
-                watchers.notifyMsg(msg.latestProcessedMsgId)
-            }
-
-            // previously I errored here, but we need to just ignore them -
-            // the transition proc submits a BlockUploaded as part of finishing the BlockBoundary messages.
-            is ReplicaMessage.BlockUploaded -> {
-                watchers.notifyMsg(msg.latestProcessedMsgId)
-            }
-
-            is ReplicaMessage.NoOp -> msg.srcMsgId?.let { watchers.notifyMsg(it) }
-
-            is ReplicaMessage.TriesDeleted -> {
-                trieCatalog.deleteTries(fromSchemaAndTable(msg.tableName), msg.trieKeys)
-            }
-        }
-    }
-
-    override suspend fun processRecords(records: List<Log.Record<ReplicaMessage>>) {
-        LOG.debug("[$dbName] transition: processing ${records.size} records")
-
-        for (record in records) {
-            val msgId = record.msgId
-            LOG.trace { "[$dbName] transition: message $msgId (${record.message::class.simpleName})" }
-
-            try {
-                if (!record.message.stale) processRecord(record)
-
-                latestReplicaMsgId = msgId
-            } catch (e: InterruptedException) {
-                throw e
-            } catch (e: Interrupted) {
-                throw e
-            } catch (e: Throwable) {
-                LOG.error(
-                    e,
-                    "[$dbName] transition: failed to process log record with msgId $msgId (${record.message::class.simpleName})"
-                )
-                watchers.notifyError(e)
-                throw e
-            }
-        }
-    }
-
-    override fun close() {
-        allocator.close()
-    }
-}
